Split HobbyChips into various files
