+++ client/flutter/fadzmaq/lib/views/widgets/matchEntry.dart
@@ -2,7 +2,7 @@ import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/views/profilepage.dart';
 import 'package:fadzmaq/views/widgets/displayPhoto.dart';
-import 'package:fadzmaq/views/widgets/hobbyChip.dart';
+import 'package:fadzmaq/views/widgets/hobbyChips.dart';
 import 'package:flutter/material.dart';
 
 class MatchEntry extends StatelessWidget {
