+++ client/flutter/fadzmaq/lib/views/widgets/profile_body.dart
@@ -2,7 +2,7 @@ import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/hobbies.dart';
 import 'package:fadzmaq/models/profile.dart';
 import 'package:fadzmaq/controllers/profile.dart';
-import 'package:fadzmaq/views/widgets/hobbyChip.dart';
+import 'package:fadzmaq/views/widgets/hobbyChips.dart';
 import 'package:flutter/material.dart';
 
 /// Helper (method?) to the ProfileFieldWidget.
